requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2
sqlparse==0.4.2
whitenoise==6.2.0
django-crispy-forms==2.0
django-allauth==0.54.0
django-jazzmin==2.6.0
crispy_bootstrap5
django-jazzmin